12.10.0 commit

+++ mk/spksrc.wheel.mk
@@ -62,11 +62,15 @@ build_wheel_target: $(PRE_WHEEL_TARGET)
 		$(foreach e,$(shell cat $(WORK_DIR)/python-cc.mk),$(eval $(e))) \
 		if [ ! -z "$(CROSS_COMPILE_WHEELS)" ] ; then \
 			$(MSG) "Force cross-compile" ; \
-			$(RUN) CFLAGS="$(CFLAGS) -I$(STAGING_INSTALL_PREFIX)/$(PYTHON_INC_DIR) $(WHEELS_CFLAGS)" LDFLAGS="$(LDFLAGS) $(WHEELS_LDFLAGS)" $(PIP_WHEEL) ; \
+			if [ -z "$(CROSSENV)" ]; then \
+				$(RUN) _PYTHON_HOST_PLATFORM="$(TC_TARGET)" CFLAGS="$(CFLAGS) -I$(STAGING_INSTALL_PREFIX)/$(PYTHON_INC_DIR) $(WHEELS_CFLAGS)" LDFLAGS="$(LDFLAGS) $(WHEELS_LDFLAGS)" $(PIP_WHEEL) ; \
+			else \
+				. $(CROSSENV) && $(RUN) _PYTHON_HOST_PLATFORM="$(TC_TARGET)" CFLAGS="$(CFLAGS) -I$(STAGING_INSTALL_PREFIX)/$(PYTHON_INC_DIR) $(WHEELS_CFLAGS)" LDFLAGS="$(LDFLAGS) $(WHEELS_LDFLAGS)" pip $(PIP_WHEEL_ARGS) --no-build-isolation ; \
+			fi ; \
 		else \
 			$(MSG) "Force pure-python" ; \
 			export LD= LDSHARED= CPP= NM= CC= AS= RANLIB= CXX= AR= STRIP= OBJDUMP= READELF= CFLAGS= CPPFLAGS= CXXFLAGS= LDFLAGS= && \
-			  $(RUN) $(PIP_WHEEL) ; \
+				$(RUN) $(PIP_WHEEL) ; \
 		fi ; \
 	fi
 
@@ -74,10 +78,20 @@ build_wheel_target: $(PRE_WHEEL_TARGET)
 post_wheel_target: $(WHEEL_TARGET)
 	@if [ -d "$(WORK_DIR)/wheelhouse" ] ; then \
 		mkdir -p $(STAGING_INSTALL_PREFIX)/share/wheelhouse ; \
-		cd $(WORK_DIR)/wheelhouse && \
-		  for w in *.whl; do \
-		    cp -f $$w $(STAGING_INSTALL_PREFIX)/share/wheelhouse/`echo $$w | cut -d"-" -f -3`-none-any.whl; \
-		  done ; \
+		cd $(WORK_DIR)/wheelhouse ; \
+		cp requirements.txt $(STAGING_INSTALL_PREFIX)/share/wheelhouse/ ; \
+		if [ "$(EXCLUDE_PURE_PYTHON_WHEELS)" = "yes" ] ; then \
+			echo "Pure python wheels are excluded from the package wheelhouse." ; \
+			for w in *.whl; do \
+				if echo $${w} | grep -viq "-none-any\.whl" ; then \
+					cp -f $$w $(STAGING_INSTALL_PREFIX)/share/wheelhouse/`echo $$w | cut -d"-" -f -3`-none-any.whl; \
+				fi ; \
+			done ; \
+		else \
+			for w in *.whl; do \
+				cp -f $$w $(STAGING_INSTALL_PREFIX)/share/wheelhouse/`echo $$w | cut -d"-" -f -3`-none-any.whl; \
+			done ; \
+		fi ; \
 	fi
